Moving LT(0, x) overrides to keymap level
